+++ src/components/App/App.js
@@ -9,19 +9,22 @@ import {
 // TODO: Figure out why this alias doesn't work
 import menu from '../../config/menu.json';
 import AppRoutes from '../../routes';
+import NotificationProvider from 'components/Notification/Notification.provider';
 import './App.scss';
 
 const App = () => (
   <Router>
-    <div className="jumpstart-app-container">
-      <NavBar menu={menu} />
-      <div className="jumpstart-app-content">
-        <AppRoutes />
+    <NotificationProvider>
+      <div className="jumpstart-app-container">
+        <NavBar menu={menu} />
+        <div className="jumpstart-app-content">
+          <AppRoutes />
+        </div>
+        <DebugLogger />
+        <ModalContainer />
+        <NotificationContainer />
       </div>
-      <DebugLogger />
-      <ModalContainer />
-      <NotificationContainer />
-    </div>
+    </NotificationProvider>
   </Router>
 );