+++ src/components/Notification/Notification.context.js
@@ -1,0 +1,5 @@
+import React,{ useContext } from 'react';
+
+export const NotificationContext = React.createContext([]);
+
+export const useNotificationValue = () => useContext(NotificationContext);