+++ src/state/notificationState.js
@@ -1,1 +1,14 @@
-export default [];
+export default [
+  {
+    type: 'info',
+    title: 'This is a notification',
+    message: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+    id: 98
+  },
+  {
+    type: 'danger',
+    title: 'This is a notification',
+    message: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+    id: 99
+  }
+];